+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Show warnings from the component registry during uploading components
 - Add config for ReadTheDocs
 - Add `dest-dir` option to `compote component pack` command to specify the destination directory for the archive
+- Add `compote registry login` CLI command to login to the component registry and store the token in the config file
 
 ### Fixed